requirements.txt
# ============================================
# کتابخانه‌های اصلی علم داده
# ============================================
pandas==2.0.3              # کار با دیتافریم
numpy==1.24.3             # محاسبات عددی
scipy==1.11.1            # محاسبات آماری

# ============================================
# تحلیل سری‌های زمانی
# ============================================
statsmodels==0.14.0       # ARIMA, SARIMA, decomposition
prophet==1.1.5           # مدل Prophet فیسبوک
pmdarima==2.0.4          # Auto ARIMA
sktime==0.24.0           # ابزارهای سری زمانی

# ============================================
# یادگیری ماشین
# ============================================
scikit-learn==1.3.0       # preprocessing, metrics

# ============================================
# مصورسازی
# ============================================
matplotlib==3.7.2         # رسم نمودار
seaborn==0.12.2          # نمودارهای آماری
plotly==5.17.0           # نمودارهای تعاملی
plotly-express==0.4.1    # مصورسازی سریع

# ============================================
# رابط کاربری
# ============================================
streamlit==1.28.1        # برنامه وب

# ============================================
# ابزارهای کمکی
# ============================================
joblib==1.3.2           # ذخیره مدل
pyyaml==6.0.1           # فایل تنظیمات
python-dotenv==1.0.0    # متغیرهای محیطی

# ============================================
# توسعه و تست
# ============================================
jupyter==1.0.0          # نوتبوک
pytest==7.4.0           # تست‌نویسی
black==23.7.0           # فرمت کد
flake8==6.1.0           # لینت کد